Add get-in-touch page and eslint
- apply eslint to codebase
- add github workflow for checking eslint
- separate header and footer
- make the background blur on HeaderOverlay active
- add standard .eslintrc
- lint existing code with eslint
