+++ .eslintrc
@@ -1,0 +1,9 @@
+{
+  "root": true,
+  "extends": ["@nuxt/eslint-config"],
+  "rules": {
+    "vue/multi-word-component-names": "off",
+    "vue/no-v-html": "off",
+    "no-useless-escape": "off"
+  }
+}
