+++ components/HeaderOverlay.vue
@@ -1,18 +1,21 @@
 <template>
   <div
-    class="fixed top-0 left-0 w-full h-full bg-white opacity-99 overflow-x-auto "
+    class="fixed top-0 right-0 w-full h-full bg-white opacity-99 overflow-x-auto "
   >
     <div class="flex py-16 px-8 justify-between">
-      <img src="~/assets/icons/logo.svg" />
-      <div @click="emit('close')" class=" bg-sky-500">
+      <img src="~/assets/icons/logo.svg">
+      <div
+        class=" bg-sky-500"
+        @click="emit('close')"
+      >
         <div class="p-1 rounded-md cursor-pointer close-hover">
-          <img src="~/assets/icons/close.svg" />
+          <img src="~/assets/icons/close.svg">
         </div>
       </div>
     </div>
     <div class="relative">
       <div class="absolute">
-        <div class="bg-gradient-blue"></div>
+        <div class="bg-gradient-blue" />
       </div>
     </div>
     <div
@@ -21,19 +24,21 @@
       class="py-8 px-8 border-b border-primary cursor-pointer"
       @click="emitScroll(header.scrollTo, header.text)"
     >
-      <h3 class="text-[28px] font-semibold">{{ header.text }}</h3>
+      <h3 class="text-[28px] font-semibold">
+        {{ header.text }}
+      </h3>
     </div>
     <div class="flex justify-between items-center py-[32px]">
       <div>
         <a href="https://calendly.com/ashwath-2/30min?month=2023-06">
           <div class="mx-8 bg-secondary text-white py-1.5 px-3 rounded-lg text-lg">
-          Book a Demo
-        </div>
+            Book a Demo
+          </div>
         </a>
       </div>
       <div>
         <div style="height: 80px; width: 100%">
-          <img src="~/assets/icons/dots-overlay.svg" />
+          <img src="~/assets/icons/dots-overlay.svg">
         </div>
       </div>
     </div>
@@ -66,12 +71,14 @@ function emitScroll(scrollTo, headerText) {
   position: absolute;
   top: -130px;
 }
+
 .background-gradient {
   position: absolute;
   top: 108px;
 }
-.close-hover:hover{
-background: #F5F6F6;
-border-radius: 5px;
+
+.close-hover:hover {
+  background: #F5F6F6;
+  border-radius: 5px;
 }
 </style>
